+++ srcs/ft_converters_1.c
@@ -6,7 +6,7 @@
 /*   By: bamrouch <bamrouch@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/06 18:02:26 by bamrouch          #+#    #+#             */
-/*   Updated: 2022/11/09 00:42:16 by bamrouch         ###   ########.fr       */
+/*   Updated: 2022/11/10 00:52:31 by bamrouch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,12 +30,12 @@ size_t	ft_print_str(char *str)
 	return (i);
 }
 
-size_t	ft_print_pointer(unsigned int p)
+size_t	ft_print_pointer(unsigned long long p)
 {
 	return (ft_print_str("0x") + ft_print_hex(p, 0));
 }
 
-size_t	ft_print_number(int i)
+size_t	ft_print_number(long long i)
 {
 	return (ft_print_decimal(i));
 }
